A few changes made.
Created a displayMenu() function in main.cpp, to allow the user to select which option they would like to see from the menu. I also deleted unnecessary comments of my other attempts to load in the data from the text file and populate the vector. I have started to implement the code for isWayBlocked(), after I get this working the next thing to do is implement the logic for the move() function.