Displaying all cells working.
I created a list of pairs to hold the 10 x 10 board and populated it. I checked if each cell matches the position of a bug and displayed them based on this. Need to figure out how to remove the comma at the end of a line in this method, and for my display history method.